Organize getFunctions
